Make glue and lal Recommended, but not hard dependencies
lal is used only for LIGOTimeGPS, and the import is at the point of
use and not at module level. glue is used to provide old ILWD format
support, and glue imports are guarded by if statements. The basic
functionality of python-ligo-lw is still present if lal and glue are
missing, so it is appopriate to describe the relationship using
`Recommends` rather than `Depends`.

On the other hand, LALSuite uses glue pervasively, so LALSuite must
Depend on python-ligo-lw.

This breaks the circular dependency that was a problem in
lscsoft/lalsuite!644.